+++ config/database.php
@@ -58,7 +58,12 @@ return [
             'sslmode' => 'prefer',
         ],
 
-        
+        'testing' => [
+            'driver' => 'sqlite',
+            'database' => ':memory:',
+            'prefix' => '',
+            'foreign_key_constraints' => env('DB_FOREIGN_KEYS', true),
+        ],
 
     ],